requirements.txt
streamlit
streamlit-folium
folium
requests
supabase
reportlab
Pillow
